+++ packages/terra-slide-group/src/terra-dev-site/test/slide-group/OneItemSlideGroup.test.jsx
@@ -1,9 +1,13 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import SlideGroup from '../../../SlideGroup';
+import styles from './DefaultOneItemSlideGroup.test.scss';
+
+const cx = classNames.bind(styles);
 
 export default () => (
   <SlideGroup
-    style={{ height: '200px' }}
+    className={cx('content-wrapper')}
     items={[
       <div className="slide1" key="1">Slide 1</div>,
     ]}