+++ packages/terra-slide-panel/src/terra-dev-site/test/slide-panel/SlidePanelSquishLarge.test.jsx
@@ -1,11 +1,15 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import SlidePanel from '../../../SlidePanel';
+import styles from './SlidePanelSquishLarge.test.scss';
+
+const cx = classNames.bind(styles);
 
 const SlidePanelSquishLarge = () => (
-  <div style={{ height: '100vh' }}>
+  <div className={cx('content-wrapper')}>
     <SlidePanel
-      mainContent={<div style={{ height: '100%', width: '100%', backgroundColor: 'red' }} />}
-      panelContent={<div style={{ height: '100%', width: '100%', backgroundColor: 'blue' }} />}
+      mainContent={<div className={cx('main-content')} />}
+      panelContent={<div className={cx('panel-content')} />}
       panelBehavior="squish"
       panelSize="large"
       isOpen